changed GET getSubcategories() to GET getCategory, renamed exceptions package, fixed NumberFormatException throwing
