+++ src/main/java/com/example/springlabs/exception/CategoryNotFoundException.java
@@ -1,8 +1,12 @@
-package com.example.springlabs.exceptions;
+package com.example.springlabs.exception;
 
 public class CategoryNotFoundException extends RuntimeException {
 
   public CategoryNotFoundException(String message) {
     super(message);
+  }
+
+  public CategoryNotFoundException(String message, Throwable cause) {
+    super(message, cause);
   }
 }
