+++ src/main/java/com/example/springlabs/services/CategoryService.java
@@ -1,16 +1,15 @@
 package com.example.springlabs.services;
 
-import com.example.springlabs.exceptions.CategoryNotFoundException;
+import com.example.springlabs.exception.CategoryNotFoundException;
 import com.example.springlabs.model.Category;
 import com.example.springlabs.repositories.CategoryRepository;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Optional;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import org.springframework.stereotype.Service;
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,7 +19,7 @@ public class CategoryService {
   CategoryRepository categoryRepository;
 
   String CATEGORY_NAME_NOT_FOUND = "Category with name: %s doesn't exist!";
-  String CATEGORY_ID_NOT_FOUND = "Category with id: %d doesn't exist!";
+  String CATEGORY_ID_NOT_FOUND = "Category with id: %s doesn't exist!";
 
   public List<Category> getAllCategories() {
     return categoryRepository.getCategories();
@@ -36,26 +35,9 @@ public class CategoryService {
     return categoryFromDto;
   }
 
-  public Collection<Category> getSubcategories(List<String> categoriesNames) {
-    if (!categoriesNames.isEmpty()) {
-      Optional<Category> parentCategory = categoryRepository.getCategoryByName(
-          categoriesNames.get(0));
-      for (int i = 1; i < categoriesNames.size(); i++) {
-        if (parentCategory.isEmpty()) {
-          throw new CategoryNotFoundException(
-              CATEGORY_NAME_NOT_FOUND.formatted(categoriesNames.get(i - 1)));
-        }
-        parentCategory = getSubcategoryByName(parentCategory.get(), categoriesNames.get(i));
-      }
-
-      return parentCategory.orElseThrow(
-              () -> new CategoryNotFoundException(
-                  CATEGORY_NAME_NOT_FOUND.formatted(categoriesNames.get(categoriesNames.size() - 1))))
-          .getSubCategories();
+    public Collection<Category> getSubcategories(List<String> categoriesNames) {
+        return getCategoryByName(categoriesNames).getSubCategories();
     }
-
-    return Collections.emptyList();
-  }
 
   private Optional<Category> getSubcategoryByName(Category category, String name) {
     return category.getSubCategories().stream()
@@ -63,14 +45,39 @@ public class CategoryService {
         .findFirst();
   }
 
-  public Category updateCategory(List<String> categoriesNames, long id,
-      Category categoryFromDto) {
-      return categoryRepository.updateCategory(id, categoryFromDto, getSubcategories(categoriesNames))
-          .orElseThrow(() -> new CategoryNotFoundException(CATEGORY_ID_NOT_FOUND.formatted(id)));
-  }
+    public Category getCategoryByName(List<String> categoriesNamesPath) {
+        Optional<Category> parentCategory = categoryRepository.getCategoryByName(categoriesNamesPath.get(0));
+        for (int i = 1; i < categoriesNamesPath.size(); i++) {
+            if (parentCategory.isEmpty()) {
+                throw new CategoryNotFoundException(
+                        CATEGORY_NAME_NOT_FOUND.formatted(categoriesNamesPath.get(i - 1)));
+            }
+            parentCategory = getSubcategoryByName(parentCategory.get(), categoriesNamesPath.get(i));
+        }
 
-  public void deleteCategory(List<String> categoriesNames, long id) {
-    categoryRepository.deleteCategory(id, getSubcategories(categoriesNames));
+        return parentCategory.orElseThrow(
+                () -> new CategoryNotFoundException(
+                        CATEGORY_NAME_NOT_FOUND.formatted(categoriesNamesPath.get(categoriesNamesPath.size() - 1))));
+    }
+
+    public Category updateCategory(List<String> categoriesNames, String stringId,
+                                   Category categoryFromDto) {
+        try {
+            long id = Long.parseLong(stringId);
+            return categoryRepository.updateCategory(id, categoryFromDto, getSubcategories(categoriesNames))
+                    .orElseThrow(() -> new CategoryNotFoundException(CATEGORY_ID_NOT_FOUND.formatted(stringId)));
+        } catch (NumberFormatException e) {
+            throw new CategoryNotFoundException(CATEGORY_ID_NOT_FOUND.formatted(stringId), e);
+        }
+    }
+
+  public void deleteCategory(List<String> categoriesNames, String stringId) {
+      try {
+          long id = Long.parseLong(stringId);
+          categoryRepository.deleteCategory(id, getSubcategories(categoriesNames));
+      } catch (NumberFormatException e) {
+          throw new CategoryNotFoundException(CATEGORY_ID_NOT_FOUND.formatted(stringId), e);
+      }
   }
 }
